chore(gate): lint formal/ — it was gate-EXECUTED but never gate-LINTED
formal/gate.sh runs mutate.py --check-anchors, openapi_conformance.py
--strict and 850 differential tests out of formal/, but its ruff phase
read src/ tests/ only. `ruff check formal/` reported 177 errors. Same
shape as two defects found the same day: tui/ outside the ruff gate hid a
log-pane regression, and a lane no CI path ran hid an identity-keyed cache
bug. A directory the gate EXECUTES but does not LINT is a blind spot.

formal/ now joins the ruff scope in all three places that lint: gate.sh
phase (c), scripts/pre_commit.sh, and lint-gate.yml (command AND path
filters). 177 -> 0.

The class that mattered was 14 unused locals / unpacked variables / loop
variables. In a differential harness that is not cosmetic: an unused
`lean_*` binding means both sides were computed and only one compared.
Three were exactly that, and are now wired:

  * bank_selection: the hoard scenario dropped `lean_crit`. Criticality is
    NOT an uncovered property — five siblings compare it, including the
    240-example randomized test — but this scenario compared nothing.
  * bank_selection: `_assert_freeze` left uncalled in the tasks-coin test.
    The author `_`-prefixed what they meant to drop (`_pc`/`_lc`) and left
    `state, gd` unprefixed.
  * equipment_scoring: `meta` carries `cur_stats` per slot precisely so the
    WORN item's score can be checked, and it never was — Lean's `cur_score`
    entered only as a lower bound whenever the feasible argmax won.

Each added assertion was broken on purpose and observed to FAIL before
being kept, so none is decorative. The remaining eleven were leftovers:
`_`-prefixed (the file convention: a prefix is a DECISION, no prefix is a
SIGNAL) or, where a sibling test already covers the property, deleted with
a comment naming the covering test.

mutate.py got extra care — its mutant payloads must match production
source character for character. Six long lines were reflowed with implicit
string concatenation only (a compile-time identity), verified twice: a
sha256 over every module-level constant is unchanged
(36b733fb194cc18e...), and --check-anchors reports 785 mutations, all
resolving uniquely, 0 via re-indent normalisation. No anchor broke.
generate_lean_fixture.py was likewise run before and after; its
286,121-char output is byte-identical.

No blanket formal/ exclusion — that would recreate the gap. Four targeted
suppressions, each with its reason at the site: E701 on cycle_step.py
(fires_mirror is a column-aligned table mirroring ProductionLadder.fires
arm-for-arm), RUF005 on two 46-int positional oracle layouts, SIM103 on a
guard chain, E501 on one injected line of production source.

Also raised four oracle request/response zips from ruff's `strict=False`
to `strict=True`: a length mismatch there silently skips comparisons,
which is the same defect class this commit audits.

Gate: bash formal/gate.sh rc=0 — ruff clean, 850 differential, 7797 + 266
python, 100.00% coverage.

+++ formal/diff/test_action_cost_nonneg_diff.py
@@ -15,7 +15,8 @@ Strategy:
 * Per-action structural assertions pin the formula.
 * Cross-check Lean oracle on the four structural cores.
 """
-from hypothesis import given, settings, strategies as st
+from hypothesis import given, settings
+from hypothesis import strategies as st
 
 from artifactsmmo_cli.ai.actions.accept_task import AcceptTaskAction
 from artifactsmmo_cli.ai.actions.claim import ClaimPendingItemAction
@@ -156,7 +157,7 @@ def test_rest_cost_pure_nonneg(max_hp, frac):
     Also spot-checks the formula, which is denominated in REAL SECONDS: a full
     deficit (hp=0) is 100.0, a full-HP rest is 3.0 (the 3s minimum), and a
     10%-missing deficit is 10.0."""
-    hp = int(round((1.0 - frac) * max_hp))
+    hp = round((1.0 - frac) * max_hp)
     hp = max(0, min(hp, max_hp))
     out = rest_cost_pure(hp, max_hp)
     assert out >= 0.0
@@ -190,7 +191,7 @@ def test_overheal_sentinel_dominates_rest_cost(max_hp, frac):
 
     Agreement of the two constants is asserted separately against the Lean value
     itself, in `test_overheal_sentinel_matches_lean`."""
-    hp = int(round((1.0 - frac) * max_hp))
+    hp = round((1.0 - frac) * max_hp)
     hp = max(0, min(hp, max_hp))
     assert rest_cost_pure(hp, max_hp) < OVERHEAL_CONSUMABLE_COST
     # The bound is tight at a full deficit: 100.0 < 200.0, with REST_COST_MAX the
